Update HydroEOS.py
Rolled back a change that made the highP ice EOS too small (length P_MPa==10), which led to ignored curvature of the profiles for large moons, leading to a step-like profile. Similar changes were made around February 2024 in other files such as Geophysical.py that set various EOS to size 10, as part of a commit with the goal of preserving memory usage. Work is in progress to see if those also need to be changed back.

Signed-off-by: Steve Vance <vance.steven@gmail.com>