Update to make CustomSolution capability compatible with Inductogram. Specifically, reworks how EOS file generation is done to prevent simultaneous writing in parallel planet runs and reworks how w_ppt input is used.
For w_ppt, if the user inputs a value other than 0.0, PlanetProfile proportionally adjusts the CustomSolution input until it matches the input w_ppt. Otherwise of 0.0 is input, then PlanetProfile assumes the input CustomSolution string is the absolute ratio and does not adjust it. This makes w_ppt a usable parameter to query over using the Inductogram.

Also adds a quick fix to get multiple CustomSolution runs to use different color schemes. Added this so I can generate multi-model figures for poster where different CustomSolution runs have different colors. This is hard coded in and does not allow for flexibility in the colors used. Should be revisited in future to make it more consistent with PlanetProfile coding structure.