Bug fixing updates:
Adds some bug fixing updates that are related to when we re-calculate an EOS. Particularly relevant for multiple runs of parallel computing when different cores calculate an EOS in one parallel run but then in another run the planet is handed off to another core. Ultimately, this leads to errros where we think we have the EOS loaded but we do not. Added checks to recalculate the EOS in this case.

In the future, we should look into a shared dictionary or simpler solution to share EOS across cores so we prevent having to re calculate EOS, which take up a majority of runtime.